+++ sdk/go/azure/insights/v20150401/pulumiTypes.go
@@ -866,6 +866,8 @@ func (o EmailNotificationResponsePtrOutput) SendToSubscriptionCoAdministrators()
 type MetricTrigger struct {
 	// List of dimension conditions. For example: [{"DimensionName":"AppName","Operator":"Equals","Values":["App1"]},{"DimensionName":"Deployment","Operator":"Equals","Values":["default"]}].
 	Dimensions []ScaleRuleMetricDimension `pulumi:"dimensions"`
+	// a value indicating whether metric should divide per instance.
+	DividePerInstance *bool `pulumi:"dividePerInstance"`
 	// the name of the metric that defines what the rule monitors.
 	MetricName string `pulumi:"metricName"`
 	// the namespace of the metric that defines what the rule monitors.
@@ -901,6 +903,8 @@ type MetricTriggerInput interface {
 type MetricTriggerArgs struct {
 	// List of dimension conditions. For example: [{"DimensionName":"AppName","Operator":"Equals","Values":["App1"]},{"DimensionName":"Deployment","Operator":"Equals","Values":["default"]}].
 	Dimensions ScaleRuleMetricDimensionArrayInput `pulumi:"dimensions"`
+	// a value indicating whether metric should divide per instance.
+	DividePerInstance pulumi.BoolPtrInput `pulumi:"dividePerInstance"`
 	// the name of the metric that defines what the rule monitors.
 	MetricName pulumi.StringInput `pulumi:"metricName"`
 	// the namespace of the metric that defines what the rule monitors.
@@ -953,6 +957,11 @@ func (o MetricTriggerOutput) Dimensions() ScaleRuleMetricDimensionArrayOutput {
 	return o.ApplyT(func(v MetricTrigger) []ScaleRuleMetricDimension { return v.Dimensions }).(ScaleRuleMetricDimensionArrayOutput)
 }
 
+// a value indicating whether metric should divide per instance.
+func (o MetricTriggerOutput) DividePerInstance() pulumi.BoolPtrOutput {
+	return o.ApplyT(func(v MetricTrigger) *bool { return v.DividePerInstance }).(pulumi.BoolPtrOutput)
+}
+
 // the name of the metric that defines what the rule monitors.
 func (o MetricTriggerOutput) MetricName() pulumi.StringOutput {
 	return o.ApplyT(func(v MetricTrigger) string { return v.MetricName }).(pulumi.StringOutput)
@@ -1002,6 +1011,8 @@ func (o MetricTriggerOutput) TimeWindow() pulumi.StringOutput {
 type MetricTriggerResponse struct {
 	// List of dimension conditions. For example: [{"DimensionName":"AppName","Operator":"Equals","Values":["App1"]},{"DimensionName":"Deployment","Operator":"Equals","Values":["default"]}].
 	Dimensions []ScaleRuleMetricDimensionResponse `pulumi:"dimensions"`
+	// a value indicating whether metric should divide per instance.
+	DividePerInstance *bool `pulumi:"dividePerInstance"`
 	// the name of the metric that defines what the rule monitors.
 	MetricName string `pulumi:"metricName"`
 	// the namespace of the metric that defines what the rule monitors.
@@ -1037,6 +1048,8 @@ type MetricTriggerResponseInput interface {
 type MetricTriggerResponseArgs struct {
 	// List of dimension conditions. For example: [{"DimensionName":"AppName","Operator":"Equals","Values":["App1"]},{"DimensionName":"Deployment","Operator":"Equals","Values":["default"]}].
 	Dimensions ScaleRuleMetricDimensionResponseArrayInput `pulumi:"dimensions"`
+	// a value indicating whether metric should divide per instance.
+	DividePerInstance pulumi.BoolPtrInput `pulumi:"dividePerInstance"`
 	// the name of the metric that defines what the rule monitors.
 	MetricName pulumi.StringInput `pulumi:"metricName"`
 	// the namespace of the metric that defines what the rule monitors.
@@ -1087,6 +1100,11 @@ func (o MetricTriggerResponseOutput) ToMetricTriggerResponseOutputWithContext(ct
 // List of dimension conditions. For example: [{"DimensionName":"AppName","Operator":"Equals","Values":["App1"]},{"DimensionName":"Deployment","Operator":"Equals","Values":["default"]}].
 func (o MetricTriggerResponseOutput) Dimensions() ScaleRuleMetricDimensionResponseArrayOutput {
 	return o.ApplyT(func(v MetricTriggerResponse) []ScaleRuleMetricDimensionResponse { return v.Dimensions }).(ScaleRuleMetricDimensionResponseArrayOutput)
+}
+
+// a value indicating whether metric should divide per instance.
+func (o MetricTriggerResponseOutput) DividePerInstance() pulumi.BoolPtrOutput {
+	return o.ApplyT(func(v MetricTriggerResponse) *bool { return v.DividePerInstance }).(pulumi.BoolPtrOutput)
 }
 
 // the name of the metric that defines what the rule monitors.